@@ -1,16 +1,19 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
+
     <groupId>net.sprd.image</groupId>
     <artifactId>sprd-webp</artifactId>
-    <version>1.0.3</version>
+    <version>1.0.3-EDEV</version>
     <packaging>jar</packaging>
+
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
     </properties>
-    
+
     <licenses>
         <license>
             <name>Apache License, Version 2.0</name>
@@ -18,14 +21,13 @@
             <distribution>repo</distribution>
         </license>
     </licenses>
-    
+
     <scm>
         <url>https://github.com/spreadshirt/sprd-webp</url>
         <connection>scm:git:git@github.com:spreadshirt/sprd-webp.git</connection>
         <tag>sprd-webp-1.0.3</tag>
-    </scm>    
-    
-   
+    </scm>
+
     <dependencies>
         <dependency>
             <groupId>com.google.webp</groupId>
@@ -35,11 +37,10 @@
         <dependency>
             <groupId>org.scijava</groupId>
             <artifactId>native-lib-loader</artifactId>
-            <version>2.3.2</version>
+            <version>2.4.0</version>
         </dependency>
     </dependencies>
-    
-    
+
     <build>
         <plugins>
             <plugin>
@@ -58,23 +59,22 @@
             </plugin>
             <plugin>
                 <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.8.0</version>
+                <version>3.10.1</version>
                 <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
+                    <source>${maven.compiler.source}</source>
+                    <target>${maven.compiler.target}</target>
                     <encoding>${project.build.sourceEncoding}</encoding>
                 </configuration>
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-javadoc-plugin</artifactId>
-                <version>3.1.0</version>
-                <configuration> 
-                    <failOnError>false</failOnError>  
-                    <additionalparam>-Xdoclint:none</additionalparam> 
-                </configuration> 
+                <version>3.4.1</version>
+                <configuration>
+                    <failOnError>false</failOnError>
+                    <additionalOptions>-Xdoclint:none</additionalOptions>
+                </configuration>
             </plugin>
         </plugins>
     </build>
-    
 </project>
